Helpers for example raw data

A wrapper on top of system.file to ease Momit examples. See examples below.

example_data(x)

example_dir()

Arguments

x
character file(s) names from the example folder

Value

example_data will return the full path where passed files are; example_data will return the containing folder where examples data have been copied on your computer.

Details

Most of Momit examples use this to cope with different file locations between your machine and mine. Providing that your working directory contains the file(s) you are interested in, or if you provide argument where to from/harvest everything should work fine, just like in the vignette.

References

Adapted from R-pkg by Hadley Wickham.

Examples

 # Example folder path
 example_dir()

#> [1] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata"

 # List all example data
 list.files(example_dir(), recursive=TRUE, full.names=TRUE)

#>  [1] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/3parts.mom"              
#>  [2] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite.mom"            
#>  [3] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/amrut.mom"      
#>  [4] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/ballantines.mom"
#>  [5] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/brahma.mom"     
#>  [6] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/bushmills.mom"  
#>  [7] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/caney.mom"      
#>  [8] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/chimay.mom"     
#>  [9] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/chivas.mom"     
#> [10] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/corona.mom"     
#> [11] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/dalmore.mom"    
#> [12] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite/deusventrue.mom"
#> [13] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/mini.mom"                
#> [14] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/mom_files/foo.mom"       
#> [15] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/Momit_bot_lite.mom"      
#> [16] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/Momit_threeparts.mom"    
#> [17] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/olea.mom"                
#> [18] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/tpsDig_allowen1.tps"     
#> [19] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/tpsDig_allowen2.tps"     
#> [20] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/tpsDig_allowen3.tps"     
#> [21] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/tpsDig_allowen4.tps"     

 # path to one of them
 example_data("bot_lite.mom")

#> [1] "/Users/vbonhomme/Research/MomX/Momit/inst/extdata/bot_lite.mom"