Import morphometric files data

Import morphometric files data

from_mom(pattern = "mom$", where = getwd(), encoding = "unknown", ...)

from_tps(pattern = "tps$", where = getwd(), encoding = "unknown", ...)

Arguments

pattern
a regular expression (see regex and examples below) to subset the list of harvested files. Defaults to filetype$ for each from_filetype (eg tps$ for from_tps), etc. $ is the base::regex marker for "end of string".

where
either path(s) or a list of characters typically obtained with harvest or the raw readLines. By default, setwd() so that all files somewhere in your current working directory will be harvested and imported.

encoding
see harvest

...
additional parameters to feed harvest (and list.files in the end)

Value

mom_df that can be manipulated and exported

Details

from_* are nothing but shortcuts on top of harvest and parse. Type for example from_mom (no brackets) to see it.

Note

If you are not yet familiar with regular expression, go learn them, they are among the most profitable investment you can make when talking to a computer.

Currently supported formats

.mom with from_mom

.tps with from_tps

Examples

# See ?example_data (you should not need the `where` argument below)

# grab all files containing `mom` in the directory
# of course we have some NAs below since we are merging "different" .mom
# with different cov, etc. components
from_mom("mom", example_dir())

#> > some components were not found everywhere:
#> 	coo
#> 	cur
#> 	domes
#> 	id
#> 	ind
#> 	ldk
#> 	out
#> 	size
#> 	sp
#> 	taxa
#> 	type
#> 	var
#> 	view
#> 

#> # A tibble: 49 x 14
#>    name  id    sp    taxa  out   ldk   cur   type  size  coo   var   domes view 
#>    <chr> <chr> <chr> <chr> <lis> <lis> <lis> <chr> <chr> <lis> <chr> <chr> <chr>
#>  1 plop  1     plip  plup  <coo> <coo> <coo> NA    NA    <NUL… NA    NA    NA   
#>  2 plup  2     plup  plup  <coo> <coo> <coo> NA    NA    <NUL… NA    NA    NA   
#>  3 brah… NA    NA    NA    <NUL… <NUL… <NUL… 2     333.… <coo> NA    NA    NA   
#>  4 caney NA    NA    NA    <NUL… <NUL… <NUL… 2     306.… <coo> NA    NA    NA   
#>  5 chim… NA    NA    NA    <NUL… <NUL… <NUL… 2     220.… <coo> NA    NA    NA   
#>  6 coro… NA    NA    NA    <NUL… <NUL… <NUL… 2     244.… <coo> NA    NA    NA   
#>  7 deus… NA    NA    NA    <NUL… <NUL… <NUL… 2     279.… <coo> NA    NA    NA   
#>  8 amrut NA    NA    NA    <NUL… <NUL… <NUL… 1     266.… <coo> NA    NA    NA   
#>  9 ball… NA    NA    NA    <NUL… <NUL… <NUL… 1     248.… <coo> NA    NA    NA   
#> 10 bush… NA    NA    NA    <NUL… <NUL… <NUL… 1     275.… <coo> NA    NA    NA   
#> # … with 39 more rows, and 1 more variable: ind <chr>

# grab all files within the 'bot_lite'  folder
# this one works like a charm
from_mom("bot_lite/", example_dir())

#> # A tibble: 10 x 4
#>    name        type  size             coo   
#>    <chr>       <chr> <chr>            <list>
#>  1 amrut       1     266.024847921009 <coo> 
#>  2 ballantines 1     248.353306740193 <coo> 
#>  3 brahma      2     333.794465725393 <coo> 
#>  4 bushmills   1     275.325270268995 <coo> 
#>  5 caney       2     306.484290886628 <coo> 
#>  6 chimay      2     220.98557993466  <coo> 
#>  7 chivas      1     268.609789073204 <coo> 
#>  8 corona      2     244.834254654269 <coo> 
#>  9 dalmore     1     234.419221812847 <coo> 
#> 10 deusventrue 2     279.25940916688  <coo> 

# same for tps, all files that begin with tpsDig_allo
from_tps("tpsDig_allo", example_dir())

#> > some components were not found everywhere:
#> 	POINTS_1
#> 	POINTS_2
#> 	POINTS_3
#> 	POINTS_4
#> 

#> # A tibble: 8 x 8
#>   name             ID    SCALE    LM     POINTS_1 POINTS_2 POINTS_3 POINTS_4
#>   <chr>            <chr> <chr>    <list> <list>   <list>   <list>   <list>  
#> 1 nhm-11684R-4.jpg 0     0.003890 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 2 nhm-17357R-4.jpg 2     0.003662 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 3 nhm-11684R-4.jpg 0     0.003890 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 4 nhm-17357R-4.jpg 2     0.003662 <coo>  <coo>    <coo>    <coo>    <NULL>  
#> 5 nhm-11684R-4.jpg 0     0.003890 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 6 nhm-17357R-4.jpg 2     0.003662 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 7 nhm-11684R-4.jpg 0     0.003890 <coo>  <NULL>   <NULL>   <NULL>   <NULL>  
#> 8 nhm-17357R-4.jpg 2     0.003662 <coo>  <coo>    <coo>    <coo>    <coo>   

# more restrictively
from_tps("tpsDig_allowen1", example_dir())

#> # A tibble: 2 x 8
#>   name             ID    SCALE    LM     POINTS_1 POINTS_2 POINTS_3 POINTS_4
#>   <chr>            <chr> <chr>    <list> <list>   <list>   <list>   <list>  
#> 1 nhm-11684R-4.jpg 0     0.003890 <coo>  <coo>    <coo>    <coo>    <coo>   
#> 2 nhm-17357R-4.jpg 2     0.003662 <coo>  <coo>    <coo>    <coo>    <coo>   

# or if you prefer a real path rather than a matching pattern
# (again, you should not need 'example_data()'), see ?example_data
harvest(example_data("tpsDig_allowen1.tps")) %>%
  parse_tps()

#> Error in harvest(example_data("tpsDig_allowen1.tps")): 
#> x all files were found empty