Detect and read files

A convenient wrapper around list.files and readLines that also prepare files for the momification.

harvest(pattern = "", where = getwd(), encoding = "UTF-8", ...)

Arguments

pattern
a regex to feed list.files (default to `` no subsetting)

where
character a list of folders or files (default getwd())

encoding
one of UTF-8 (default) or latin1

...
more arguments to feed list.files

Value

a list of character lines

Note

from_* importers harvest internally so most of the time you should not need harvest alone. Internally, list.files is passed with full.names=TRUE, recursive = TRUE, so that full names are returned and all subfolders (if any) will be scrapped. If you do not want this behaviour, you can still harvest directly after the proper list (or single) of paths, typically obtained with list.files.

Also, some problems with encoding may occur, particularly with latin1. from_* importers, when used directly, should message you with "see ?harvest". In that case, try changing the encoding parameter (see readLines).

Examples

# See ?example_data (you should not need the `where` argument below)

# grab all files containing `mom` in the directory
harvest("mom", example_dir())

#> [[1]]
#>  [1] "~plop"              "id 1 "              "sp plip"           
#>  [4] "taxa plup"          "out"                " 697  977"         
#>  [7] "766        991    " "704 1046"           "629 1008 "         
#> [10] "731  926"           "541  962"           "835  911"          
#> [13] "ldk"                "541 962"            "542 965"           
#> [16] "592 1073"           "593 1075"           "cur"               
#> [19] "541 949"            "542 952"            "542 954"           
#> [22] ""                   "~plup"              "id 2 "             
#> [25] "sp plup"            "taxa plup"          "out"               
#> [28] " 693  974"          "762        993 "    "709 1043"          
#> [31] "624 1009 "          "732  921"           "543  960"          
#> [34] "837  912"           "ldk"                "541 964"           
#> [37] "532 985"            "594 1073"           "593 1075"          
#> [40] "cur"                "541 949"            "541 953"           
#> [43] "546 954"           
#> 
#> [[2]]
#>   [1] "~brahma"               "type 2"                "size 333.794465725393"
#>   [4] "37 561"                "63 361"                "42 150"               
#>   [7] "143 15"                "316 113"               "305 302"              
#>  [10] "295 523"               "296 713"               "240 912"              
#>  [13] "205 1101"              "110 1002"              "82 781"               
#>  [16] "~caney"                "type 2"                "size 306.484290886628"
#>  [19] "53 535"                "53 343"                "53 162"               
#>  [22] "144 28"                "304 99"                "311 291"              
#>  [25] "311 462"               "272 644"               "246 834"              
#>  [28] "235 1009"              "129 927"               "111 746"              
#>  [31] "~chimay"               "type 2"                "size 220.98557993466" 
#>  [34] "49 333"                "52 202"                "56 79"                
#>  [37] "148 13"                "276 29"                "291 148"              
#>  [40] "288 279"               "258 416"               "212 537"              
#>  [43] "194 654"               "114 598"               "98 478"               
#>  [46] "~corona"               "type 2"                "size 244.834254654269"
#>  [49] "91 426"                "91 271"                "94 128"               
#>  [52] "164 17"                "275 80"                "276 225"              
#>  [55] "276 380"               "246 532"               "236 672"              
#>  [58] "216 822"               "144 735"               "124 597"              
#>  [61] "~deusventrue"          "type 2"                "size 279.25940916688" 
#>  [64] "74 481"                "36 310"                "46 147"               
#>  [67] "160 54"                "305 97"                "319 272"              
#>  [70] "308 434"               "242 604"               "225 775"              
#>  [73] "211 932"               "130 828"               "131 678"              
#>  [76] "~amrut"                "type 1"                "size 266.024847921009"
#>  [79] "57 441"                "57 281"                "57 112"               
#>  [82] "129 11"                "268 64"                "268 224"              
#>  [85] "268 383"               "252 553"               "208 704"              
#>  [88] "193 871"               "122 784"               "111 616"              
#>  [91] "~ballantines"          "type 1"                "size 248.353306740193"
#>  [94] "38 357"                "38 215"                "36 74"                
#>  [97] "142 3"                 "284 4"                 "311 138"              
#> [100] "310 296"               "307 437"               "238 558"              
#> [103] "208 701"               "132 643"               "51 522"               
#> [106] "~bushmills"            "type 1"                "size 275.325270268995"
#> [109] "72 456"                "75 287"                "77 133"               
#> [112] "146 11"                "291 60"                "284 229"              
#> [115] "287 399"               "290 568"               "253 718"              
#> [118] "221 881"               "132 805"               "68 656"               
#> [121] "~chivas"               "type 1"                "size 268.609789073204"
#> [124] "33 437"                "32 273"                "36 125"               
#> [127] "141 44"                "297 52"                "329 201"              
#> [130] "332 349"               "273 508"               "238 668"              
#> [133] "232 816"               "130 770"               "113 614"              
#> [136] "~dalmore"              "type 1"                "size 234.419221812847"
#> [139] "52 371"                "42 235"                "40 87"                
#> [142] "117 34"                "252 35"                "315 112"              
#> [145] "312 260"               "286 396"               "219 529"              
#> [148] "212 665"               "135 677"               "130 528"              
#> 
#> [[3]]
#>  [1] "~amrut"                "type 1"                "size 266.024847921009"
#>  [4] "57 441"                "57 281"                "57 112"               
#>  [7] "129 11"                "268 64"                "268 224"              
#> [10] "268 383"               "252 553"               "208 704"              
#> [13] "193 871"               "122 784"               "111 616"              
#> 
#> [[4]]
#>  [1] "~ballantines"          "type 1"                "size 248.353306740193"
#>  [4] "38 357"                "38 215"                "36 74"                
#>  [7] "142 3"                 "284 4"                 "311 138"              
#> [10] "310 296"               "307 437"               "238 558"              
#> [13] "208 701"               "132 643"               "51 522"               
#> 
#> [[5]]
#>  [1] "~brahma"               "type 2"                "size 333.794465725393"
#>  [4] "37 561"                "63 361"                "42 150"               
#>  [7] "143 15"                "316 113"               "305 302"              
#> [10] "295 523"               "296 713"               "240 912"              
#> [13] "205 1101"              "110 1002"              "82 781"               
#> 
#> [[6]]
#>  [1] "~bushmills"            "type 1"                "size 275.325270268995"
#>  [4] "72 456"                "75 287"                "77 133"               
#>  [7] "146 11"                "291 60"                "284 229"              
#> [10] "287 399"               "290 568"               "253 718"              
#> [13] "221 881"               "132 805"               "68 656"               
#> 
#> [[7]]
#>  [1] "~caney"                "type 2"                "size 306.484290886628"
#>  [4] "53 535"                "53 343"                "53 162"               
#>  [7] "144 28"                "304 99"                "311 291"              
#> [10] "311 462"               "272 644"               "246 834"              
#> [13] "235 1009"              "129 927"               "111 746"              
#> 
#> [[8]]
#>  [1] "~chimay"              "type 2"               "size 220.98557993466"
#>  [4] "49 333"               "52 202"               "56 79"               
#>  [7] "148 13"               "276 29"               "291 148"             
#> [10] "288 279"              "258 416"              "212 537"             
#> [13] "194 654"              "114 598"              "98 478"              
#> 
#> [[9]]
#>  [1] "~chivas"               "type 1"                "size 268.609789073204"
#>  [4] "33 437"                "32 273"                "36 125"               
#>  [7] "141 44"                "297 52"                "329 201"              
#> [10] "332 349"               "273 508"               "238 668"              
#> [13] "232 816"               "130 770"               "113 614"              
#> 
#> [[10]]
#>  [1] "~corona"               "type 2"                "size 244.834254654269"
#>  [4] "91 426"                "91 271"                "94 128"               
#>  [7] "164 17"                "275 80"                "276 225"              
#> [10] "276 380"               "246 532"               "236 672"              
#> [13] "216 822"               "144 735"               "124 597"              
#> 
#> [[11]]
#>  [1] "~dalmore"              "type 1"                "size 234.419221812847"
#>  [4] "52 371"                "42 235"                "40 87"                
#>  [7] "117 34"                "252 35"                "315 112"              
#> [10] "312 260"               "286 396"               "219 529"              
#> [13] "212 665"               "135 677"               "130 528"              
#> 
#> [[12]]
#>  [1] "~deusventrue"         "type 2"               "size 279.25940916688"
#>  [4] "74 481"               "36 310"               "46 147"              
#>  [7] "160 54"               "305 97"               "319 272"             
#> [10] "308 434"              "242 604"              "225 775"             
#> [13] "211 932"              "130 828"              "131 678"             
#> 
#> [[13]]
#> [1] "766 991 " "704 1046" "629 1008" "731  926"
#> 
#> [[14]]
#>  [1] "~0001-cPastor_O10VD"   "var Pastor"            "domes cult"           
#>  [4] "view VD"               "ind O10"               "size 2.23"            
#>  [7] "-0.5000 0.00000"       "-0.4967 0.01035"       "-0.4935 0.02414"      
#> [10] "~0001-cPastor_O10VL"   "var Pastor"            "domes cult"           
#> [13] "view VL"               "ind O10"               "size 2.43"            
#> [16] "-0.5000 0.00000"       "-0.4995 0.01018"       "-0.4957 0.02022"      
#> [19] "~0001-cIvorra_O11VL"   "var Ivorra"            "domes wild"           
#> [22] "view VL"               "ind O11"               "size 2.56"            
#> [25] "-0.5000 0.00000"       "-0.4985 0.0112"        "-0.4845 0.021"        
#> [28] "~0001-cIvorra_O11VD"   "var Ivorra"            "domes wild"           
#> [31] "view VD"               "ind O11"               "size 2.74"            
#> [34] "-0.5000 0.00000"       "-0.4995 0.011"         "-0.49 0.0198"         
#> [37] "~0005-cBonhomme_O11VD" "var Bonhomme"          "domes wild"           
#> [40] "view VD"               "ind O11"               "size 2.55"            
#> [43] "-0.5000 0.00000"       "-0.499 0.012"          "-0.49 0.0186"         
#> 
#> [[15]]
#>   [1] "~brahma"               "type 2"                "size 333.794465725393"
#>   [4] "37 561"                "63 361"                "42 150"               
#>   [7] "143 15"                "316 113"               "305 302"              
#>  [10] "295 523"               "296 713"               "240 912"              
#>  [13] "205 1101"              "110 1002"              "82 781"               
#>  [16] "~caney"                "type 2"                "size 306.484290886628"
#>  [19] "53 535"                "53 343"                "53 162"               
#>  [22] "144 28"                "304 99"                "311 291"              
#>  [25] "311 462"               "272 644"               "246 834"              
#>  [28] "235 1009"              "129 927"               "111 746"              
#>  [31] "~chimay"               "type 2"                "size 220.98557993466" 
#>  [34] "49 333"                "52 202"                "56 79"                
#>  [37] "148 13"                "276 29"                "291 148"              
#>  [40] "288 279"               "258 416"               "212 537"              
#>  [43] "194 654"               "114 598"               "98 478"               
#>  [46] "~corona"               "type 2"                "size 244.834254654269"
#>  [49] "91 426"                "91 271"                "94 128"               
#>  [52] "164 17"                "275 80"                "276 225"              
#>  [55] "276 380"               "246 532"               "236 672"              
#>  [58] "216 822"               "144 735"               "124 597"              
#>  [61] "~deusventrue"          "type 2"                "size 279.25940916688" 
#>  [64] "74 481"                "36 310"                "46 147"               
#>  [67] "160 54"                "305 97"                "319 272"              
#>  [70] "308 434"               "242 604"               "225 775"              
#>  [73] "211 932"               "130 828"               "131 678"              
#>  [76] "~amrut"                "type 1"                "size 266.024847921009"
#>  [79] "57 441"                "57 281"                "57 112"               
#>  [82] "129 11"                "268 64"                "268 224"              
#>  [85] "268 383"               "252 553"               "208 704"              
#>  [88] "193 871"               "122 784"               "111 616"              
#>  [91] "~ballantines"          "type 1"                "size 248.353306740193"
#>  [94] "38 357"                "38 215"                "36 74"                
#>  [97] "142 3"                 "284 4"                 "311 138"              
#> [100] "310 296"               "307 437"               "238 558"              
#> [103] "208 701"               "132 643"               "51 522"               
#> [106] "~bushmills"            "type 1"                "size 275.325270268995"
#> [109] "72 456"                "75 287"                "77 133"               
#> [112] "146 11"                "291 60"                "284 229"              
#> [115] "287 399"               "290 568"               "253 718"              
#> [118] "221 881"               "132 805"               "68 656"               
#> [121] "~chivas"               "type 1"                "size 268.609789073204"
#> [124] "33 437"                "32 273"                "36 125"               
#> [127] "141 44"                "297 52"                "329 201"              
#> [130] "332 349"               "273 508"               "238 668"              
#> [133] "232 816"               "130 770"               "113 614"              
#> [136] "~dalmore"              "type 1"                "size 234.419221812847"
#> [139] "52 371"                "42 235"                "40 87"                
#> [142] "117 34"                "252 35"                "315 112"              
#> [145] "312 260"               "286 396"               "219 529"              
#> [148] "212 665"               "135 677"               "130 528"              
#> 
#> [[16]]
#>  [1] "~plop"              "id 1 "              "sp plip"           
#>  [4] "taxa plup"          "out"                " 697  977"         
#>  [7] "766        991    " "704 1046"           "629 1008 "         
#> [10] "731  926"           "541  962"           "835  911"          
#> [13] "ldk"                "541 962"            "542 965"           
#> [16] "592 1073"           "593 1075"           "cur"               
#> [19] "541 949"            "542 952"            "542 954"           
#> [22] ""                   "~plup"              "id 2 "             
#> [25] "sp plup"            "taxa plup"          "out"               
#> [28] " 693  974"          "762        993 "    "709 1043"          
#> [31] "624 1009 "          "732  921"           "543  960"          
#> [34] "837  912"           "ldk"                "541 964"           
#> [37] "532 985"            "594 1073"           "593 1075"          
#> [40] "cur"                "541 949"            "541 953"           
#> [43] "546 954"           
#> 
#> [[17]]
#>   [1] "~0001-cAglan_O10VD" "var Aglan"          "domes cult"        
#>   [4] "view VD"            "ind O10"            "-0.5 0"            
#>   [7] "-0.4531 0.08999"    "-0.3718 0.177"      "-0.2697 0.2472"    
#>  [10] "-0.1774 0.2828"     "-0.08494 0.3047"    "0.007775 0.3094"   
#>  [13] "0.1006 0.3072"      "0.1936 0.2879"      "0.287 0.2445"      
#>  [16] "0.3739 0.177"       "0.4438 0.09201"     "~0001-cAglan_O10VL"
#>  [19] "var Aglan"          "domes cult"         "view VL"           
#>  [22] "ind O10"            "-0.5 0"             "-0.4589 0.08322"   
#>  [25] "-0.3709 0.1508"     "-0.2768 0.201"      "-0.1872 0.2277"    
#>  [28] "-0.0948 0.2406"     "-0.01658 0.2405"    "0.07184 0.24"      
#>  [31] "0.163 0.2258"       "0.2503 0.1981"      "0.3397 0.1431"     
#>  [34] "0.4251 0.07474"     "~0001-cAglan_O11VD" "var Aglan"         
#>  [37] "domes cult"         "view VD"            "ind O11"           
#>  [40] "-0.5 0"             "-0.4287 0.08482"    "-0.3635 0.1634"    
#>  [43] "-0.2887 0.1982"     "-0.1984 0.2394"     "-0.1047 0.2462"    
#>  [46] "-0.0141 0.2562"     "0.07666 0.2349"     "0.1674 0.2104"     
#>  [49] "0.2614 0.1766"      "0.3398 0.1302"      "0.4276 0.07767"    
#>  [52] "~0001-cAglan_O11VL" "var Aglan"          "domes cult"        
#>  [55] "view VL"            "ind O11"            "-0.5 0"            
#>  [58] "-0.4613 0.1004"     "-0.4112 0.1785"     "-0.3182 0.2485"    
#>  [61] "-0.2297 0.2878"     "-0.1294 0.3141"     "-0.03993 0.3068"   
#>  [64] "0.06116 0.2835"     "0.1589 0.2542"      "0.2469 0.2161"     
#>  [67] "0.3462 0.1557"      "0.433 0.0898"       "~0001-cAglan_O12VD"
#>  [70] "var Aglan"          "domes cult"         "view VD"           
#>  [73] "ind O12"            "-0.5 0"             "-0.4447 0.08998"   
#>  [76] "-0.3676 0.1833"     "-0.2834 0.2585"     "-0.1888 0.2973"    
#>  [79] "-0.07988 0.3179"    "0.01435 0.3205"     "0.1085 0.3158"     
#>  [82] "0.2169 0.2893"      "0.3107 0.2521"      "0.4006 0.1895"     
#>  [85] "0.4757 0.09447"     "~0001-cAglan_O12VL" "var Aglan"         
#>  [88] "domes cult"         "view VL"            "ind O12"           
#>  [91] "-0.5 0"             "-0.4483 0.09002"    "-0.3716 0.1435"    
#>  [94] "-0.277 0.186"       "-0.1862 0.2067"     "-0.09177 0.2238"   
#>  [97] "-0.001122 0.2301"   "0.0822 0.2292"      "0.1726 0.2101"     
#> [100] "0.2664 0.1728"      "0.3384 0.1286"      "0.4209 0.05883"    
#> [103] "~0001-cAglan_O13VD" "var Aglan"          "domes cult"        
#> [106] "view VD"            "ind O13"            "-0.5 0"            
#> [109] "-0.4218 0.06662"    "-0.344 0.1395"      "-0.2555 0.1878"    
#> [112] "-0.1661 0.2174"     "-0.07586 0.2313"    "0.0145 0.2421"     
#> [115] "0.1177 0.2473"      "0.2062 0.2329"      "0.2978 0.2187"     
#> [118] "0.4005 0.1706"      "0.4736 0.08958"     "~0001-cAglan_O13VL"
#> [121] "var Aglan"          "domes cult"         "view VL"           
#> [124] "ind O13"            "-0.5 0"             "-0.4387 0.0718"    
#> [127] "-0.3568 0.1151"     "-0.263 0.1486"      "-0.1729 0.1638"    
#> [130] "-0.09269 0.1639"    "-0.003196 0.1637"   "0.0863 0.1635"     
#> [133] "0.1664 0.1605"      "0.2583 0.1417"      "0.3433 0.1015"     
#> [136] "0.4187 0.05548"     "~0001-cAglan_O14VD" "var Aglan"         
#> [139] "domes cult"         "view VD"            "ind O14"           
#> [142] "-0.5 0"             "-0.4556 0.09085"    "-0.3721 0.187"     
#> [145] "-0.2836 0.2507"     "-0.1873 0.2885"     "-0.09606 0.2999"   
#> [148] "0.005301 0.3052"    "0.0941 0.3"         "0.1868 0.2819"     
#> [151] "0.291 0.2349"       "0.3729 0.1737"      "0.4465 0.08244"    
#> [154] "~0001-cAglan_O14VL" "var Aglan"          "domes cult"        
#> [157] "view VL"            "ind O14"            "-0.5 0"            
#> [160] "-0.4652 0.102"      "-0.4174 0.2032"     "-0.3278 0.2957"    
#> [163] "-0.2313 0.3391"     "-0.1263 0.3592"     "-0.03511 0.364"    
#> [166] "0.06794 0.3485"     "0.1672 0.3236"      "0.2686 0.279"      
#> [169] "0.3651 0.2023"      "0.4376 0.1041"     
#> 

# same for tps, all files that have the .tps extension
harvest("\\.tps$", example_dir())

#> [[1]]
#>   [1] "LM=12"                  "1404.00000 707.00000"   "1325.00000 841.00000"  
#>   [4] "1316.00000 604.00000"   "1233.00000 691.00000"   "1085.00000 695.00000"  
#>   [7] "988.00000 834.00000"    "1014.00000 645.00000"   "848.00000 709.00000"   
#>  [10] "938.00000 955.00000"    "994.00000 508.00000"    "1319.00000 916.00000"  
#>  [13] "1281.00000 494.00000"   "CURVES=4"               "POINTS=15"             
#>  [16] "993.00000 507.00000"    "1014.00000 505.00000"   "1034.00000 496.00000"  
#>  [19] "1055.00000 488.00000"   "1074.00000 479.00000"   "1093.00000 469.00000"  
#>  [22] "1114.00000 462.00000"   "1135.00000 459.00000"   "1157.00000 456.00000"  
#>  [25] "1178.00000 459.00000"   "1199.00000 462.00000"   "1220.00000 469.00000"  
#>  [28] "1241.00000 476.00000"   "1261.00000 484.00000"   "1280.00000 494.00000"  
#>  [31] "POINTS=30"              "1281.00000 492.00000"   "1328.00000 485.00000"  
#>  [34] "1374.00000 473.00000"   "1418.00000 457.00000"   "1465.00000 453.00000"  
#>  [37] "1513.00000 459.00000"   "1558.00000 473.00000"   "1602.00000 490.00000"  
#>  [40] "1649.00000 497.00000"   "1695.00000 508.00000"   "1738.00000 527.00000"  
#>  [43] "1784.00000 538.00000"   "1825.00000 561.00000"   "1855.00000 597.00000"  
#>  [46] "1869.00000 642.00000"   "1870.00000 690.00000"   "1864.00000 736.00000"  
#>  [49] "1843.00000 778.00000"   "1812.00000 814.00000"   "1780.00000 849.00000"  
#>  [52] "1738.00000 870.00000"   "1692.00000 880.00000"   "1645.00000 883.00000"  
#>  [55] "1597.00000 886.00000"   "1552.00000 900.00000"   "1505.00000 903.00000"  
#>  [58] "1459.00000 915.00000"   "1412.00000 920.00000"   "1367.00000 908.00000"  
#>  [61] "1320.00000 915.00000"   "POINTS=20"              "1320.00000 917.00000"  
#>  [64] "1301.00000 927.00000"   "1282.00000 937.00000"   "1264.00000 948.00000"  
#>  [67] "1247.00000 960.00000"   "1228.00000 970.00000"   "1209.00000 979.00000"  
#>  [70] "1189.00000 986.00000"   "1168.00000 988.00000"   "1146.00000 989.00000"  
#>  [73] "1125.00000 987.00000"   "1104.00000 983.00000"   "1084.00000 978.00000"  
#>  [76] "1063.00000 972.00000"   "1043.00000 966.00000"   "1022.00000 961.00000"  
#>  [79] "1002.00000 956.00000"   "980.00000 954.00000"    "959.00000 953.00000"   
#>  [82] "938.00000 955.00000"    "POINTS=30"              "937.00000 956.00000"   
#>  [85] "901.00000 969.00000"    "865.00000 982.00000"    "828.00000 994.00000"   
#>  [88] "790.00000 997.00000"    "752.00000 994.00000"    "715.00000 985.00000"   
#>  [91] "681.00000 968.00000"    "655.00000 940.00000"    "637.00000 906.00000"   
#>  [94] "630.00000 868.00000"    "646.00000 834.00000"    "662.00000 799.00000"   
#>  [97] "663.00000 761.00000"    "654.00000 724.00000"    "652.00000 686.00000"   
#> [100] "641.00000 649.00000"    "634.00000 612.00000"    "647.00000 576.00000"   
#> [103] "665.00000 542.00000"    "686.00000 510.00000"    "706.00000 477.00000"   
#> [106] "735.00000 453.00000"    "773.00000 449.00000"    "811.00000 452.00000"   
#> [109] "849.00000 458.00000"    "886.00000 466.00000"    "922.00000 480.00000"   
#> [112] "956.00000 498.00000"    "993.00000 508.00000"    "IMAGE=nhm-11684R-4.jpg"
#> [115] "ID=0 "                  "SCALE=0.003890"         "LM=12"                 
#> [118] "1379.00000 879.00000"   "1285.00000 1019.00000"  "1279.00000 759.00000"  
#> [121] "1204.00000 882.00000"   "1022.00000 891.00000"   "912.00000 1043.00000"  
#> [124] "962.00000 811.00000"    "772.00000 896.00000"    "832.00000 1118.00000"  
#> [127] "858.00000 661.00000"    "1275.00000 1081.00000"  "1272.00000 687.00000"  
#> [130] "CURVES=4"               "POINTS=15"              "860.00000 662.00000"   
#> [133] "890.00000 658.00000"    "921.00000 654.00000"    "951.00000 650.00000"   
#> [136] "982.00000 645.00000"    "1011.00000 637.00000"   "1041.00000 631.00000"  
#> [139] "1072.00000 627.00000"   "1102.00000 624.00000"   "1133.00000 624.00000"  
#> [142] "1163.00000 632.00000"   "1191.00000 644.00000"   "1218.00000 658.00000"  
#> [145] "1246.00000 672.00000"   "1273.00000 686.00000"   "POINTS=30"             
#> [148] "1271.00000 687.00000"   "1316.00000 677.00000"   "1360.00000 662.00000"  
#> [151] "1407.00000 660.00000"   "1454.00000 666.00000"   "1501.00000 672.00000"  
#> [154] "1546.00000 686.00000"   "1589.00000 706.00000"   "1635.00000 718.00000"  
#> [157] "1682.00000 723.00000"   "1728.00000 734.00000"   "1769.00000 757.00000"  
#> [160] "1803.00000 790.00000"   "1826.00000 831.00000"   "1849.00000 872.00000"  
#> [163] "1855.00000 918.00000"   "1841.00000 963.00000"   "1815.00000 1002.00000" 
#> [166] "1773.00000 1023.00000"  "1727.00000 1016.00000"  "1683.00000 1016.00000" 
#> [169] "1645.00000 1045.00000"  "1603.00000 1068.00000"  "1559.00000 1082.00000" 
#> [172] "1512.00000 1088.00000"  "1464.00000 1089.00000"  "1417.00000 1092.00000" 
#> [175] "1369.00000 1091.00000"  "1323.00000 1083.00000"  "1277.00000 1082.00000" 
#> [178] "POINTS=20"              "1275.00000 1083.00000"  "1255.00000 1097.00000" 
#> [181] "1235.00000 1111.00000"  "1214.00000 1123.00000"  "1192.00000 1134.00000" 
#> [184] "1170.00000 1142.00000"  "1146.00000 1147.00000"  "1122.00000 1149.00000" 
#> [187] "1097.00000 1148.00000"  "1073.00000 1146.00000"  "1049.00000 1142.00000" 
#> [190] "1025.00000 1136.00000"  "1002.00000 1130.00000"  "978.00000 1127.00000"  
#> [193] "954.00000 1124.00000"   "929.00000 1122.00000"   "905.00000 1121.00000"  
#> [196] "881.00000 1120.00000"   "856.00000 1120.00000"   "830.00000 1119.00000"  
#> [199] "POINTS=30"              "832.00000 1119.00000"   "799.00000 1134.00000"  
#> [202] "765.00000 1147.00000"   "729.00000 1154.00000"   "693.00000 1157.00000"  
#> [205] "656.00000 1157.00000"   "621.00000 1149.00000"   "591.00000 1129.00000"  
#> [208] "576.00000 1096.00000"   "569.00000 1061.00000"   "559.00000 1026.00000"  
#> [211] "575.00000 994.00000"    "589.00000 961.00000"    "599.00000 926.00000"   
#> [214] "612.00000 892.00000"    "621.00000 857.00000"    "618.00000 821.00000"   
#> [217] "606.00000 786.00000"    "586.00000 756.00000"    "580.00000 722.00000"   
#> [220] "583.00000 686.00000"    "593.00000 651.00000"    "615.00000 623.00000"   
#> [223] "649.00000 612.00000"    "686.00000 607.00000"    "722.00000 611.00000"   
#> [226] "757.00000 619.00000"    "791.00000 632.00000"    "825.00000 646.00000"   
#> [229] "858.00000 661.00000"    "IMAGE=nhm-17357R-4.jpg" "ID=2 "                 
#> [232] "SCALE=0.003662"        
#> 
#> [[2]]
#>   [1] "LM=12"                  "1404.00000 707.00000"   "1325.00000 841.00000"  
#>   [4] "1316.00000 604.00000"   "1233.00000 691.00000"   "1085.00000 695.00000"  
#>   [7] "988.00000 834.00000"    "1014.00000 645.00000"   "848.00000 709.00000"   
#>  [10] "938.00000 955.00000"    "994.00000 508.00000"    "1319.00000 916.00000"  
#>  [13] "1281.00000 494.00000"   "CURVES=4"               "POINTS=15"             
#>  [16] "993.00000 507.00000"    "1014.00000 505.00000"   "1034.00000 496.00000"  
#>  [19] "1055.00000 488.00000"   "1074.00000 479.00000"   "1093.00000 469.00000"  
#>  [22] "1114.00000 462.00000"   "1135.00000 459.00000"   "1157.00000 456.00000"  
#>  [25] "1178.00000 459.00000"   "1199.00000 462.00000"   "1220.00000 469.00000"  
#>  [28] "1241.00000 476.00000"   "1261.00000 484.00000"   "1280.00000 494.00000"  
#>  [31] "POINTS=30"              "1281.00000 492.00000"   "1328.00000 485.00000"  
#>  [34] "1374.00000 473.00000"   "1418.00000 457.00000"   "1465.00000 453.00000"  
#>  [37] "1513.00000 459.00000"   "1558.00000 473.00000"   "1602.00000 490.00000"  
#>  [40] "1649.00000 497.00000"   "1695.00000 508.00000"   "1738.00000 527.00000"  
#>  [43] "1784.00000 538.00000"   "1825.00000 561.00000"   "1855.00000 597.00000"  
#>  [46] "1869.00000 642.00000"   "1870.00000 690.00000"   "1864.00000 736.00000"  
#>  [49] "1843.00000 778.00000"   "1812.00000 814.00000"   "1780.00000 849.00000"  
#>  [52] "1738.00000 870.00000"   "1692.00000 880.00000"   "1645.00000 883.00000"  
#>  [55] "1597.00000 886.00000"   "1552.00000 900.00000"   "1505.00000 903.00000"  
#>  [58] "1459.00000 915.00000"   "1412.00000 920.00000"   "1367.00000 908.00000"  
#>  [61] "1320.00000 915.00000"   "POINTS=20"              "1320.00000 917.00000"  
#>  [64] "1301.00000 927.00000"   "1282.00000 937.00000"   "1264.00000 948.00000"  
#>  [67] "1247.00000 960.00000"   "1228.00000 970.00000"   "1209.00000 979.00000"  
#>  [70] "1189.00000 986.00000"   "1168.00000 988.00000"   "1146.00000 989.00000"  
#>  [73] "1125.00000 987.00000"   "1104.00000 983.00000"   "1084.00000 978.00000"  
#>  [76] "1063.00000 972.00000"   "1043.00000 966.00000"   "1022.00000 961.00000"  
#>  [79] "1002.00000 956.00000"   "980.00000 954.00000"    "959.00000 953.00000"   
#>  [82] "938.00000 955.00000"    "POINTS=30"              "937.00000 956.00000"   
#>  [85] "901.00000 969.00000"    "865.00000 982.00000"    "828.00000 994.00000"   
#>  [88] "790.00000 997.00000"    "752.00000 994.00000"    "715.00000 985.00000"   
#>  [91] "681.00000 968.00000"    "655.00000 940.00000"    "637.00000 906.00000"   
#>  [94] "630.00000 868.00000"    "646.00000 834.00000"    "662.00000 799.00000"   
#>  [97] "663.00000 761.00000"    "654.00000 724.00000"    "652.00000 686.00000"   
#> [100] "641.00000 649.00000"    "634.00000 612.00000"    "647.00000 576.00000"   
#> [103] "665.00000 542.00000"    "686.00000 510.00000"    "706.00000 477.00000"   
#> [106] "735.00000 453.00000"    "773.00000 449.00000"    "811.00000 452.00000"   
#> [109] "849.00000 458.00000"    "886.00000 466.00000"    "922.00000 480.00000"   
#> [112] "956.00000 498.00000"    "993.00000 508.00000"    "IMAGE=nhm-11684R-4.jpg"
#> [115] "ID=0 "                  "SCALE=0.003890"         "LM=12"                 
#> [118] "1379.00000 879.00000"   "1285.00000 1019.00000"  "1279.00000 759.00000"  
#> [121] "1204.00000 882.00000"   "1022.00000 891.00000"   "912.00000 1043.00000"  
#> [124] "962.00000 811.00000"    "772.00000 896.00000"    "832.00000 1118.00000"  
#> [127] "858.00000 661.00000"    "1275.00000 1081.00000"  "1272.00000 687.00000"  
#> [130] "CURVES=4"               "POINTS=15"              "860.00000 662.00000"   
#> [133] "890.00000 658.00000"    "921.00000 654.00000"    "951.00000 650.00000"   
#> [136] "982.00000 645.00000"    "1011.00000 637.00000"   "1041.00000 631.00000"  
#> [139] "1072.00000 627.00000"   "1102.00000 624.00000"   "1133.00000 624.00000"  
#> [142] "1163.00000 632.00000"   "1191.00000 644.00000"   "1218.00000 658.00000"  
#> [145] "1246.00000 672.00000"   "1273.00000 686.00000"   "POINTS=20"             
#> [148] "1275.00000 1083.00000"  "1255.00000 1097.00000"  "1235.00000 1111.00000" 
#> [151] "1214.00000 1123.00000"  "1192.00000 1134.00000"  "1170.00000 1142.00000" 
#> [154] "1146.00000 1147.00000"  "1122.00000 1149.00000"  "1097.00000 1148.00000" 
#> [157] "1073.00000 1146.00000"  "1049.00000 1142.00000"  "1025.00000 1136.00000" 
#> [160] "1002.00000 1130.00000"  "978.00000 1127.00000"   "954.00000 1124.00000"  
#> [163] "929.00000 1122.00000"   "905.00000 1121.00000"   "881.00000 1120.00000"  
#> [166] "856.00000 1120.00000"   "830.00000 1119.00000"   "POINTS=30"             
#> [169] "832.00000 1119.00000"   "799.00000 1134.00000"   "765.00000 1147.00000"  
#> [172] "729.00000 1154.00000"   "693.00000 1157.00000"   "656.00000 1157.00000"  
#> [175] "621.00000 1149.00000"   "591.00000 1129.00000"   "576.00000 1096.00000"  
#> [178] "569.00000 1061.00000"   "559.00000 1026.00000"   "575.00000 994.00000"   
#> [181] "589.00000 961.00000"    "599.00000 926.00000"    "612.00000 892.00000"   
#> [184] "621.00000 857.00000"    "618.00000 821.00000"    "606.00000 786.00000"   
#> [187] "586.00000 756.00000"    "580.00000 722.00000"    "583.00000 686.00000"   
#> [190] "593.00000 651.00000"    "615.00000 623.00000"    "649.00000 612.00000"   
#> [193] "686.00000 607.00000"    "722.00000 611.00000"    "757.00000 619.00000"   
#> [196] "791.00000 632.00000"    "825.00000 646.00000"    "858.00000 661.00000"   
#> [199] "IMAGE=nhm-17357R-4.jpg" "ID=2 "                  "SCALE=0.003662"        
#> 
#> [[3]]
#>   [1] "LM=10"                  "1316.00000 604.00000"   "1233.00000 691.00000"  
#>   [4] "1085.00000 695.00000"   "988.00000 834.00000"    "1014.00000 645.00000"  
#>   [7] "848.00000 709.00000"    "938.00000 955.00000"    "994.00000 508.00000"   
#>  [10] "1319.00000 916.00000"   "1281.00000 494.00000"   "CURVES=4"              
#>  [13] "POINTS=15"              "993.00000 507.00000"    "1014.00000 505.00000"  
#>  [16] "1034.00000 496.00000"   "1055.00000 488.00000"   "1074.00000 479.00000"  
#>  [19] "1093.00000 469.00000"   "1114.00000 462.00000"   "1135.00000 459.00000"  
#>  [22] "1157.00000 456.00000"   "1178.00000 459.00000"   "1199.00000 462.00000"  
#>  [25] "1220.00000 469.00000"   "1241.00000 476.00000"   "1261.00000 484.00000"  
#>  [28] "1280.00000 494.00000"   "POINTS=30"              "1281.00000 492.00000"  
#>  [31] "1328.00000 485.00000"   "1374.00000 473.00000"   "1418.00000 457.00000"  
#>  [34] "1465.00000 453.00000"   "1513.00000 459.00000"   "1558.00000 473.00000"  
#>  [37] "1602.00000 490.00000"   "1649.00000 497.00000"   "1695.00000 508.00000"  
#>  [40] "1738.00000 527.00000"   "1784.00000 538.00000"   "1825.00000 561.00000"  
#>  [43] "1855.00000 597.00000"   "1869.00000 642.00000"   "1870.00000 690.00000"  
#>  [46] "1864.00000 736.00000"   "1843.00000 778.00000"   "1812.00000 814.00000"  
#>  [49] "1780.00000 849.00000"   "1738.00000 870.00000"   "1692.00000 880.00000"  
#>  [52] "1645.00000 883.00000"   "1597.00000 886.00000"   "1552.00000 900.00000"  
#>  [55] "1505.00000 903.00000"   "1459.00000 915.00000"   "1412.00000 920.00000"  
#>  [58] "1367.00000 908.00000"   "1320.00000 915.00000"   "POINTS=20"             
#>  [61] "1320.00000 917.00000"   "1301.00000 927.00000"   "1282.00000 937.00000"  
#>  [64] "1264.00000 948.00000"   "1247.00000 960.00000"   "1228.00000 970.00000"  
#>  [67] "1209.00000 979.00000"   "1189.00000 986.00000"   "1168.00000 988.00000"  
#>  [70] "1146.00000 989.00000"   "1125.00000 987.00000"   "1104.00000 983.00000"  
#>  [73] "1084.00000 978.00000"   "1063.00000 972.00000"   "1043.00000 966.00000"  
#>  [76] "1022.00000 961.00000"   "1002.00000 956.00000"   "980.00000 954.00000"   
#>  [79] "959.00000 953.00000"    "938.00000 955.00000"    "POINTS=30"             
#>  [82] "937.00000 956.00000"    "901.00000 969.00000"    "865.00000 982.00000"   
#>  [85] "828.00000 994.00000"    "790.00000 997.00000"    "752.00000 994.00000"   
#>  [88] "715.00000 985.00000"    "681.00000 968.00000"    "655.00000 940.00000"   
#>  [91] "637.00000 906.00000"    "630.00000 868.00000"    "646.00000 834.00000"   
#>  [94] "662.00000 799.00000"    "663.00000 761.00000"    "654.00000 724.00000"   
#>  [97] "652.00000 686.00000"    "641.00000 649.00000"    "634.00000 612.00000"   
#> [100] "647.00000 576.00000"    "665.00000 542.00000"    "686.00000 510.00000"   
#> [103] "706.00000 477.00000"    "735.00000 453.00000"    "773.00000 449.00000"   
#> [106] "811.00000 452.00000"    "849.00000 458.00000"    "886.00000 466.00000"   
#> [109] "922.00000 480.00000"    "956.00000 498.00000"    "993.00000 508.00000"   
#> [112] "IMAGE=nhm-11684R-4.jpg" "ID=0 "                  "SCALE=0.003890"        
#> [115] "LM=12"                  "1379.00000 879.00000"   "1285.00000 1019.00000" 
#> [118] "1279.00000 759.00000"   "1204.00000 882.00000"   "1022.00000 891.00000"  
#> [121] "912.00000 1043.00000"   "962.00000 811.00000"    "772.00000 896.00000"   
#> [124] "832.00000 1118.00000"   "858.00000 661.00000"    "1275.00000 1081.00000" 
#> [127] "1272.00000 687.00000"   "CURVES=4"               "POINTS=15"             
#> [130] "860.00000 662.00000"    "890.00000 658.00000"    "921.00000 654.00000"   
#> [133] "951.00000 650.00000"    "982.00000 645.00000"    "1011.00000 637.00000"  
#> [136] "1041.00000 631.00000"   "1072.00000 627.00000"   "1102.00000 624.00000"  
#> [139] "1133.00000 624.00000"   "1163.00000 632.00000"   "1191.00000 644.00000"  
#> [142] "1218.00000 658.00000"   "1246.00000 672.00000"   "1273.00000 686.00000"  
#> [145] "POINTS=30"              "1271.00000 687.00000"   "1316.00000 677.00000"  
#> [148] "1360.00000 662.00000"   "1407.00000 660.00000"   "1454.00000 666.00000"  
#> [151] "1501.00000 672.00000"   "1546.00000 686.00000"   "1589.00000 706.00000"  
#> [154] "1635.00000 718.00000"   "1682.00000 723.00000"   "1728.00000 734.00000"  
#> [157] "1769.00000 757.00000"   "1803.00000 790.00000"   "1826.00000 831.00000"  
#> [160] "1849.00000 872.00000"   "1855.00000 918.00000"   "1841.00000 963.00000"  
#> [163] "1815.00000 1002.00000"  "1773.00000 1023.00000"  "1727.00000 1016.00000" 
#> [166] "1683.00000 1016.00000"  "1645.00000 1045.00000"  "1603.00000 1068.00000" 
#> [169] "1559.00000 1082.00000"  "1512.00000 1088.00000"  "1464.00000 1089.00000" 
#> [172] "1417.00000 1092.00000"  "1369.00000 1091.00000"  "1323.00000 1083.00000" 
#> [175] "1277.00000 1082.00000"  "POINTS=20"              "1275.00000 1083.00000" 
#> [178] "1255.00000 1097.00000"  "1235.00000 1111.00000"  "1214.00000 1123.00000" 
#> [181] "1192.00000 1134.00000"  "1170.00000 1142.00000"  "1146.00000 1147.00000" 
#> [184] "1122.00000 1149.00000"  "1097.00000 1148.00000"  "1073.00000 1146.00000" 
#> [187] "1049.00000 1142.00000"  "1025.00000 1136.00000"  "1002.00000 1130.00000" 
#> [190] "978.00000 1127.00000"   "954.00000 1124.00000"   "929.00000 1122.00000"  
#> [193] "905.00000 1121.00000"   "881.00000 1120.00000"   "856.00000 1120.00000"  
#> [196] "830.00000 1119.00000"   "POINTS=30"              "832.00000 1119.00000"  
#> [199] "799.00000 1134.00000"   "765.00000 1147.00000"   "729.00000 1154.00000"  
#> [202] "693.00000 1157.00000"   "656.00000 1157.00000"   "621.00000 1149.00000"  
#> [205] "591.00000 1129.00000"   "576.00000 1096.00000"   "569.00000 1061.00000"  
#> [208] "559.00000 1026.00000"   "575.00000 994.00000"    "589.00000 961.00000"   
#> [211] "599.00000 926.00000"    "612.00000 892.00000"    "621.00000 857.00000"   
#> [214] "618.00000 821.00000"    "606.00000 786.00000"    "586.00000 756.00000"   
#> [217] "580.00000 722.00000"    "583.00000 686.00000"    "593.00000 651.00000"   
#> [220] "615.00000 623.00000"    "649.00000 612.00000"    "686.00000 607.00000"   
#> [223] "722.00000 611.00000"    "757.00000 619.00000"    "791.00000 632.00000"   
#> [226] "825.00000 646.00000"    "858.00000 661.00000"    "IMAGE=nhm-17357R-4.jpg"
#> [229] "ID=2 "                  "SCALE=0.003662"        
#> 
#> [[4]]
#>   [1] "LM=12"                  "1404.00000 707.00000"   "1325.00000 841.00000"  
#>   [4] "1316.00000 604.00000"   "1233.00000 691.00000"   "1085.00000 695.00000"  
#>   [7] "988.00000 834.00000"    "1014.00000 645.00000"   "848.00000 709.00000"   
#>  [10] "938.00000 955.00000"    "994.00000 508.00000"    "1319.00000 916.00000"  
#>  [13] "1281.00000 494.00000"   "993.00000 507.00000"    "1014.00000 505.00000"  
#>  [16] "1034.00000 496.00000"   "1055.00000 488.00000"   "1074.00000 479.00000"  
#>  [19] "1093.00000 469.00000"   "1114.00000 462.00000"   "1135.00000 459.00000"  
#>  [22] "1157.00000 456.00000"   "1178.00000 459.00000"   "1199.00000 462.00000"  
#>  [25] "1220.00000 469.00000"   "1241.00000 476.00000"   "1261.00000 484.00000"  
#>  [28] "1280.00000 494.00000"   "1281.00000 492.00000"   "1328.00000 485.00000"  
#>  [31] "1374.00000 473.00000"   "1418.00000 457.00000"   "1465.00000 453.00000"  
#>  [34] "1513.00000 459.00000"   "1558.00000 473.00000"   "1602.00000 490.00000"  
#>  [37] "1649.00000 497.00000"   "1695.00000 508.00000"   "1738.00000 527.00000"  
#>  [40] "1784.00000 538.00000"   "1825.00000 561.00000"   "1855.00000 597.00000"  
#>  [43] "1869.00000 642.00000"   "1870.00000 690.00000"   "1864.00000 736.00000"  
#>  [46] "1843.00000 778.00000"   "1812.00000 814.00000"   "1780.00000 849.00000"  
#>  [49] "1738.00000 870.00000"   "1692.00000 880.00000"   "1645.00000 883.00000"  
#>  [52] "1597.00000 886.00000"   "1552.00000 900.00000"   "1505.00000 903.00000"  
#>  [55] "1459.00000 915.00000"   "1412.00000 920.00000"   "1367.00000 908.00000"  
#>  [58] "1320.00000 915.00000"   "1320.00000 917.00000"   "1301.00000 927.00000"  
#>  [61] "1282.00000 937.00000"   "1264.00000 948.00000"   "1247.00000 960.00000"  
#>  [64] "1228.00000 970.00000"   "1209.00000 979.00000"   "1189.00000 986.00000"  
#>  [67] "1168.00000 988.00000"   "1146.00000 989.00000"   "1125.00000 987.00000"  
#>  [70] "1104.00000 983.00000"   "1084.00000 978.00000"   "1063.00000 972.00000"  
#>  [73] "1043.00000 966.00000"   "1022.00000 961.00000"   "1002.00000 956.00000"  
#>  [76] "980.00000 954.00000"    "959.00000 953.00000"    "938.00000 955.00000"   
#>  [79] "937.00000 956.00000"    "901.00000 969.00000"    "865.00000 982.00000"   
#>  [82] "828.00000 994.00000"    "790.00000 997.00000"    "752.00000 994.00000"   
#>  [85] "715.00000 985.00000"    "681.00000 968.00000"    "655.00000 940.00000"   
#>  [88] "637.00000 906.00000"    "630.00000 868.00000"    "646.00000 834.00000"   
#>  [91] "662.00000 799.00000"    "663.00000 761.00000"    "654.00000 724.00000"   
#>  [94] "652.00000 686.00000"    "641.00000 649.00000"    "634.00000 612.00000"   
#>  [97] "647.00000 576.00000"    "665.00000 542.00000"    "686.00000 510.00000"   
#> [100] "706.00000 477.00000"    "735.00000 453.00000"    "773.00000 449.00000"   
#> [103] "811.00000 452.00000"    "849.00000 458.00000"    "886.00000 466.00000"   
#> [106] "922.00000 480.00000"    "956.00000 498.00000"    "993.00000 508.00000"   
#> [109] "IMAGE=nhm-11684R-4.jpg" "ID=0 "                  "SCALE=0.003890"        
#> [112] "LM=12"                  "1379.00000 879.00000"   "1285.00000 1019.00000" 
#> [115] "1279.00000 759.00000"   "1204.00000 882.00000"   "1022.00000 891.00000"  
#> [118] "912.00000 1043.00000"   "962.00000 811.00000"    "772.00000 896.00000"   
#> [121] "832.00000 1118.00000"   "858.00000 661.00000"    "1275.00000 1081.00000" 
#> [124] "1272.00000 687.00000"   "CURVES=4"               "POINTS=15"             
#> [127] "860.00000 662.00000"    "890.00000 658.00000"    "921.00000 654.00000"   
#> [130] "951.00000 650.00000"    "982.00000 645.00000"    "1011.00000 637.00000"  
#> [133] "1041.00000 631.00000"   "1072.00000 627.00000"   "1102.00000 624.00000"  
#> [136] "1133.00000 624.00000"   "1163.00000 632.00000"   "1191.00000 644.00000"  
#> [139] "1218.00000 658.00000"   "1246.00000 672.00000"   "1273.00000 686.00000"  
#> [142] "POINTS=30"              "1271.00000 687.00000"   "1316.00000 677.00000"  
#> [145] "1360.00000 662.00000"   "1407.00000 660.00000"   "1454.00000 666.00000"  
#> [148] "1501.00000 672.00000"   "1546.00000 686.00000"   "1589.00000 706.00000"  
#> [151] "1635.00000 718.00000"   "1682.00000 723.00000"   "1728.00000 734.00000"  
#> [154] "1769.00000 757.00000"   "1803.00000 790.00000"   "1826.00000 831.00000"  
#> [157] "1849.00000 872.00000"   "1855.00000 918.00000"   "1841.00000 963.00000"  
#> [160] "1815.00000 1002.00000"  "1773.00000 1023.00000"  "1727.00000 1016.00000" 
#> [163] "1683.00000 1016.00000"  "1645.00000 1045.00000"  "1603.00000 1068.00000" 
#> [166] "1559.00000 1082.00000"  "1512.00000 1088.00000"  "1464.00000 1089.00000" 
#> [169] "1417.00000 1092.00000"  "1369.00000 1091.00000"  "1323.00000 1083.00000" 
#> [172] "1277.00000 1082.00000"  "POINTS=20"              "1275.00000 1083.00000" 
#> [175] "1255.00000 1097.00000"  "1235.00000 1111.00000"  "1214.00000 1123.00000" 
#> [178] "1192.00000 1134.00000"  "1170.00000 1142.00000"  "1146.00000 1147.00000" 
#> [181] "1122.00000 1149.00000"  "1097.00000 1148.00000"  "1073.00000 1146.00000" 
#> [184] "1049.00000 1142.00000"  "1025.00000 1136.00000"  "1002.00000 1130.00000" 
#> [187] "978.00000 1127.00000"   "954.00000 1124.00000"   "929.00000 1122.00000"  
#> [190] "905.00000 1121.00000"   "881.00000 1120.00000"   "856.00000 1120.00000"  
#> [193] "830.00000 1119.00000"   "POINTS=30"              "832.00000 1119.00000"  
#> [196] "799.00000 1134.00000"   "765.00000 1147.00000"   "729.00000 1154.00000"  
#> [199] "693.00000 1157.00000"   "656.00000 1157.00000"   "621.00000 1149.00000"  
#> [202] "591.00000 1129.00000"   "576.00000 1096.00000"   "569.00000 1061.00000"  
#> [205] "559.00000 1026.00000"   "575.00000 994.00000"    "589.00000 961.00000"   
#> [208] "599.00000 926.00000"    "612.00000 892.00000"    "621.00000 857.00000"   
#> [211] "618.00000 821.00000"    "606.00000 786.00000"    "586.00000 756.00000"   
#> [214] "580.00000 722.00000"    "583.00000 686.00000"    "593.00000 651.00000"   
#> [217] "615.00000 623.00000"    "649.00000 612.00000"    "686.00000 607.00000"   
#> [220] "722.00000 611.00000"    "757.00000 619.00000"    "791.00000 632.00000"   
#> [223] "825.00000 646.00000"    "858.00000 661.00000"    "IMAGE=nhm-17357R-4.jpg"
#> [226] "ID=2 "                  "SCALE=0.003662"        
#> 

# grab all files within the 'bot_lite'  folder
harvest("bot_lite/", example_dir())

#> [[1]]
#>  [1] "~amrut"                "type 1"                "size 266.024847921009"
#>  [4] "57 441"                "57 281"                "57 112"               
#>  [7] "129 11"                "268 64"                "268 224"              
#> [10] "268 383"               "252 553"               "208 704"              
#> [13] "193 871"               "122 784"               "111 616"              
#> 
#> [[2]]
#>  [1] "~ballantines"          "type 1"                "size 248.353306740193"
#>  [4] "38 357"                "38 215"                "36 74"                
#>  [7] "142 3"                 "284 4"                 "311 138"              
#> [10] "310 296"               "307 437"               "238 558"              
#> [13] "208 701"               "132 643"               "51 522"               
#> 
#> [[3]]
#>  [1] "~brahma"               "type 2"                "size 333.794465725393"
#>  [4] "37 561"                "63 361"                "42 150"               
#>  [7] "143 15"                "316 113"               "305 302"              
#> [10] "295 523"               "296 713"               "240 912"              
#> [13] "205 1101"              "110 1002"              "82 781"               
#> 
#> [[4]]
#>  [1] "~bushmills"            "type 1"                "size 275.325270268995"
#>  [4] "72 456"                "75 287"                "77 133"               
#>  [7] "146 11"                "291 60"                "284 229"              
#> [10] "287 399"               "290 568"               "253 718"              
#> [13] "221 881"               "132 805"               "68 656"               
#> 
#> [[5]]
#>  [1] "~caney"                "type 2"                "size 306.484290886628"
#>  [4] "53 535"                "53 343"                "53 162"               
#>  [7] "144 28"                "304 99"                "311 291"              
#> [10] "311 462"               "272 644"               "246 834"              
#> [13] "235 1009"              "129 927"               "111 746"              
#> 
#> [[6]]
#>  [1] "~chimay"              "type 2"               "size 220.98557993466"
#>  [4] "49 333"               "52 202"               "56 79"               
#>  [7] "148 13"               "276 29"               "291 148"             
#> [10] "288 279"              "258 416"              "212 537"             
#> [13] "194 654"              "114 598"              "98 478"              
#> 
#> [[7]]
#>  [1] "~chivas"               "type 1"                "size 268.609789073204"
#>  [4] "33 437"                "32 273"                "36 125"               
#>  [7] "141 44"                "297 52"                "329 201"              
#> [10] "332 349"               "273 508"               "238 668"              
#> [13] "232 816"               "130 770"               "113 614"              
#> 
#> [[8]]
#>  [1] "~corona"               "type 2"                "size 244.834254654269"
#>  [4] "91 426"                "91 271"                "94 128"               
#>  [7] "164 17"                "275 80"                "276 225"              
#> [10] "276 380"               "246 532"               "236 672"              
#> [13] "216 822"               "144 735"               "124 597"              
#> 
#> [[9]]
#>  [1] "~dalmore"              "type 1"                "size 234.419221812847"
#>  [4] "52 371"                "42 235"                "40 87"                
#>  [7] "117 34"                "252 35"                "315 112"              
#> [10] "312 260"               "286 396"               "219 529"              
#> [13] "212 665"               "135 677"               "130 528"              
#> 
#> [[10]]
#>  [1] "~deusventrue"         "type 2"               "size 279.25940916688"
#>  [4] "74 481"               "36 310"               "46 147"              
#>  [7] "160 54"               "305 97"               "319 272"             
#> [10] "308 434"              "242 604"              "225 775"             
#> [13] "211 932"              "130 828"              "131 678"             
#>