Import outline coordinates from mask image file

Given a path to an image mask, returns a [Momocs2::coo_single] with xy outline coordinates.

import_mask(x, from_col, to_col, time_limit)

Arguments

x
paths to .txt files

from_col, to_col
column names where to get paths and where to store results

time_limit
if this take more than this (in second) calculation is aborted. Useful for debugging.

Value

Momocs2::coo_single

Details

A mask (or silhouette) is ideally a grayscale image (8-bit) containing a (pure) black shape onto a (pure) white background.

This methods ensures all of this.

Calculations are delegated to algo_Conte.

.jpg files take the jpeg::readJPEG shortcut (roughly twice faster). Other files are processed with magick.

If one or more image does not work, try time_limit

See also

Examples

1+2

#> [1] 3