Example datasets and helpers

Momit_data is wraps system.file to ease Momit examples. Below is a list of datasets provided with the package and the list of the generous providers.

Momit_data(x = "")

Arguments

x
character file(s) names from the example folder, only for example_data

Value

character the full path pointing to Momit/extdata/

Details

You do not need Momit_data for your data. Most of Momit examples use these functions in examples and vignettes. For package testing, R CMD CHECK etc., it is required to find correct locations of these files.

References

Momit_data is adapted from R-pkg by Hadley Wickham.

Example files are organized in subfolders and data providers are listed below. Thanks a lot to them!

.chc file in chc/ was generated by SHAPE by Hisato Iwata and colleagues and generously shared by Callie Crawford

.tps files in tps/ were generated using tpsDig by James Rohlf and generously provided by Allowen Evin (AE/), Julien Claude and Coline Verot (JC/)

.txt files were generated using ImageJ by myself.

The vitis/ folder is intended as a minimal case of a real world database. It is at heart of the "Survival guide for handling a database" vignette (todo). jpg files in raw/ come from a huge collection of modern grapevine pips and were generously provided by Laurent Bouby, Sarah Ivorra, Vincent Bonhomme, Clémence Pagnoux, Thierry Pastor and Jean-Frédéric Terral. Three folders of this database were extracted with only 3 individuals. Image size was largely decreased to cope with CRAN standards. Structure of filenames was generously messed up. csv in meta/ is a toy example of a covariate spreadsheet.

Examples

 # List all example data
 Momit_data() %>% sniff()

#> ℹ sniffing files in /Users/runner/runners/2.169.1/work/_temp/Library/Momit/extdata/

#> # A tibble: 43 x 5
#>         size path                        path_dir                 file     ext  
#>    <fs::byt> <fs::path>                  <chr>                    <chr>    <chr>
#>  1       348 /Users/runner/runners/2.16… /Users/runner/runners/2… 3parts.… mom  
#>  2     1.25K /Users/runner/runners/2.16… /Users/runner/runners/2… Momit_b… mom  
#>  3       348 /Users/runner/runners/2.16… /Users/runner/runners/2… Momit_t… mom  
#>  4       127 /Users/runner/runners/2.16… /Users/runner/runners/2… amrut.m… mom  
#>  5       129 /Users/runner/runners/2.16… /Users/runner/runners/2… ballant… mom  
#>  6       130 /Users/runner/runners/2.16… /Users/runner/runners/2… brahma.… mom  
#>  7       130 /Users/runner/runners/2.16… /Users/runner/runners/2… bushmil… mom  
#>  8       128 /Users/runner/runners/2.16… /Users/runner/runners/2… caney.m… mom  
#>  9       125 /Users/runner/runners/2.16… /Users/runner/runners/2… chimay.… mom  
#> 10       128 /Users/runner/runners/2.16… /Users/runner/runners/2… chivas.… mom  
#> # … with 33 more rows
#> ❯mom_tbl