Helpers for example raw data

A wrapper on top of system.file to ease Momit examples. See examples below.

example_data(x)

example_dir()

Arguments

x
character file(s) names from the example folder, only for example_data

Value

character the full path pointing to the file (for example_data) or to the containing folder (for example_dir) where examples data have been copied on your computer.

Details

Most of Momit examples use these functions. This is not intended for real world, that is for your files. Here, it helps find correct locations where raw example files bundled with the package are located. Providing that your working directory contains the file(s) you are interested in, everything should work fine, just like in the vignette.

References

Adapted from R-pkg by Hadley Wickham.

Examples

 # Example folder path
 example_dir()

#> [1] "/tmp/Rtmparyg3F/temp_libpath63bf6649b8c4/Momit/extdata"

 # List all example data
 example_dir() %>% sniff()

#> ℹ sniffing files in /home/travis/build/MomX/Momit/tmp/Rtmparyg3F/temp_libpath63bf6649b8c4/Momit/extdata

#> # A tibble: 21 x 5
#>         size path                         path_dir               file      ext  
#>    <fs::byt> <fs::path>                   <chr>                  <chr>     <chr>
#>  1       348 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… 3parts.m… mom  
#>  2     1.25K /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… Momit_bo… mom  
#>  3       348 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… Momit_th… mom  
#>  4       127 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… amrut.mom mom  
#>  5       129 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… ballanti… mom  
#>  6       130 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… brahma.m… mom  
#>  7       130 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… bushmill… mom  
#>  8       128 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… caney.mom mom  
#>  9       125 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… chimay.m… mom  
#> 10       128 /tmp/Rtmparyg3F/temp_libpat… /tmp/Rtmparyg3F/temp_… chivas.m… mom  
#> # … with 11 more rows
#> ❯mom_tbl