Convert a parsed filed to a mom_df that can be manipulated and exported

Convert a parsed filed to a mom_df that can be manipulated and exported

from_mom(x, ...)

from_tps(x, ...)

from_nts(x, ...)

from_stv(x, ...)

from_lmk(x, ...)

from_StereoMorph(x, ...)

from_Optimas(x, ...)

from_PAST(x, ...)

Arguments

x
list typically what is returned by harvest; otherwise, if character will be passed to harvest with appropriate parameters. See the Momit-usage vignette.

...
additional parameters to feed harvest (and list.files in the end)

Value

mom_df that can be manipulated and exported

Note

When x is passed as a character, harvest will try to find them based on their extensions. Some format do not have their own extension (eg .txt). In the latter case and/or you should first harvest with proper parameters.

See also