Import morphometric files data

Import morphometric files data

from_mom(pattern = "mom$", x = getwd(), encoding = "unknown", ...)

from_tps(pattern = "tps$", x = getwd(), encoding = "unknown", ...)

Arguments

pattern
a regular expression (see regex and examples below) to subset the list of harvested files. Defaults to filetype$ for each from_filetype (eg tps$ for from_tps), etc.

x
either path(s) or a list of characters typically obtained with harvest or the raw readLines. By default, setwd() so that all files somewhere in your current working directory will be harvested and imported.

encoding
see harvest

...
additional parameters to feed harvest (and list.files in the end)

Value

mom_df that can be manipulated and exported

Details

from_* are nothing but shortcuts on top of harvest and parse. Type for example from_mom (no brackets) to see it.

Note

If you are not yet familiar with regular expression, go learn them, they are among the most profitable investment you can make when talking to a computer.

Currently supported formats

.mom with from_mom

.tps with from_tps